src/data-wrangling/requirements.txt
langchain-core 
langchain-community 
langchain-openai
loguru
python-dotenv
